UI: Retain submitted field values on error
If a user enters an invalid To value, then it's not helpful that both fields get reset to the default of `/`.

By keeping the submitted values, it makes it easier for the user to spot where they went wrong, and to make a fix as needed, without having to enter a (possibly very long) value again.

Fixes #43.