Tests: split into unit and integration tests
Unit tests can be run on a local host machine with `composer test`. They are quick, as they mock WP functions (via BrainMonkey) and therefore don’t need to load WordPress to work.

WP integration tests need a local server (or Travis) to be able to load WordPress and use the database. They can be run with `composer integration-test`.